# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hotel Bookings API
  version: 0.2.0
  description: REST API for the hotel-booking coding exercise.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Check service health
      operationId: healthz
      responses:
        '200':
          description: The process can accept HTTP requests.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
  /api/v1/hotels:
    get:
      summary: Find hotels by name
      operationId: searchHotels
      parameters:
        - name: name
          in: query
          required: true
          description: Case-insensitive substring to match against a hotel name.
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Matching hotels, or an empty array when none match.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hotel'
        '400':
          $ref: '#/components/responses/InvalidRequest'
  /api/v1/hotels/{hotelID}/availability:
    get:
      summary: Find rooms available for an entire stay
      operationId: findAvailableRooms
      parameters:
        - name: hotelID
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: check_in
          in: query
          required: true
          description: Inclusive ISO-8601 calendar date.
          schema:
            type: string
            format: date
            example: '2026-12-10'
        - name: check_out
          in: query
          required: true
          description: Exclusive ISO-8601 calendar date; it must be after check_in and at most 30 nights later.
          schema:
            type: string
            format: date
            example: '2026-12-12'
        - name: guests
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            example: 2
      responses:
        '200':
          description: Suitable unoccupied rooms in capacity then room-number order, or an empty array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableRoom'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          description: The hotel does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/bookings:
    post:
      summary: Create a booking and assign one suitable room
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: The booking committed and its confirmation simulation was scheduled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          description: The hotel does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: No room can accommodate the party for every requested night.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/test/reset:
    post:
      summary: Remove all evaluator test data
      description: Enabled only when ENABLE_TEST_ENDPOINTS is true. Do not expose on untrusted deployments.
      operationId: resetTestData
      responses:
        '204':
          description: All hotel, room, and booking data was removed.
        '404':
          description: Test endpoints are disabled.
  /api/v1/test/seed:
    post:
      summary: Create deterministic evaluator test data
      description: Enabled only when ENABLE_TEST_ENDPOINTS is true. Idempotently creates The Grand Hotel and its six rooms.
      operationId: seedTestData
      responses:
        '204':
          description: The deterministic test inventory exists.
        '404':
          description: Test endpoints are disabled.
components:
  schemas:
    Hotel:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: The Grand Hotel
    CreateBookingRequest:
      type: object
      additionalProperties: false
      required: [hotel_id, check_in, check_out, guest_count, lead_guest_name, lead_guest_email]
      properties:
        hotel_id:
          type: integer
          format: int64
          minimum: 1
          example: 1
        check_in:
          type: string
          format: date
          example: '2026-12-10'
        check_out:
          type: string
          format: date
          example: '2026-12-12'
        guest_count:
          type: integer
          minimum: 1
          example: 2
        lead_guest_name:
          type: string
          example: Ada Lovelace
        lead_guest_email:
          type: string
          format: email
          example: ada@example.com
    Booking:
      type: object
      required: [reference, hotel_id, room, check_in, check_out, guest_count, lead_guest_name, lead_guest_email, created_at]
      properties:
        reference:
          type: string
          pattern: '^HBK-[A-Z2-7]{12}$'
        hotel_id:
          type: integer
          format: int64
        room:
          $ref: '#/components/schemas/AvailableRoom'
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
        guest_count:
          type: integer
        lead_guest_name:
          type: string
        lead_guest_email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    AvailableRoom:
      type: object
      required: [id, number, type, capacity]
      properties:
        id:
          type: integer
          format: int64
          example: 3
        number:
          type: string
          example: '201'
        type:
          type: string
          enum: [single, double, deluxe]
        capacity:
          type: integer
          minimum: 1
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: invalid_request
            message:
              type: string
              example: name query parameter is required
  responses:
    InvalidRequest:
      description: The request is syntactically valid HTTP but violates API input requirements.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'